Add openapi and getUser by header
