+++ apps/io-func-wallet-solution-backend/src/infra/http/handlers/__test__/create-wallet-attestation.spec.ts
@@ -50,6 +50,9 @@ describe("CreateWalletAttestationHandler", async () => {
         grant_type: GRANT_TYPE_KEY_ATTESTATION,
         assertion: walletAttestationRequest,
       },
+      headers: {
+        "x-iowallet-user-id": "x-iowallet-user-id",
+      },
     };
     const handler = CreateWalletAttestationHandler({
       input: req,
