+++ apps/io-func-wallet-solution-backend/src/infra/http/handlers/create-wallet-instance.ts
@@ -1,6 +1,6 @@
 import * as t from "io-ts";
 
-import { flow, pipe } from "fp-ts/function";
+import { pipe } from "fp-ts/function";
 
 import * as H from "@pagopa/handler-kit";
 import * as RTE from "fp-ts/lib/ReaderTaskEither";
@@ -8,12 +8,12 @@ import * as E from "fp-ts/lib/Either";
 
 import { NonEmptyString } from "@pagopa/ts-commons/lib/strings";
 import { sequenceS } from "fp-ts/Apply";
-import { lookup } from "fp-ts/Record";
+
 import { logErrorAndReturnResponse } from "../utils";
 
+import { requireUser } from "./utils";
 import { insertWalletInstance, validateAttestation } from "@/wallet-instance";
 import { consumeNonce } from "@/wallet-instance-request";
-import { User } from "@/user";
 
 const WalletInstanceRequestPayload = t.type({
   challenge: NonEmptyString,
@@ -24,28 +24,6 @@ const WalletInstanceRequestPayload = t.type({
 type WalletInstanceRequestPayload = t.TypeOf<
   typeof WalletInstanceRequestPayload
 >;
-
-const requireUserId = (req: H.HttpRequest) =>
-  pipe(
-    req.headers,
-    lookup("x-iowallet-user-id"),
-    E.fromOption(
-      () => new H.HttpBadRequestError("Missing x-iowallet-user-id in header")
-    ),
-    E.chainW(
-      H.parse(
-        User.props.id,
-        "The content of x-iowallet-user-id is not a valid id"
-      )
-    )
-  );
-
-const requireUser: (
-  req: H.HttpRequest
-) => E.Either<H.HttpBadRequestError | H.ValidationError, User> = flow(
-  requireUserId,
-  E.map((id) => ({ id }))
-);
 
 const requireWalletInstanceRequest = (req: H.HttpRequest) =>
   pipe(
